fix plotserver crashing due to storing pointer to local data and update to qwt 6 api
qwt6 api requires setting interval in QtRasterData instead of using
SetBoundingRect and it wants pointers to heap objects instead of copying stack
objects.
The latter is problematic as the colormaps can not be copied in the current
design, for now they are disabled.

The QWT6 preprocessor macro is necessary so the QwtDoubleRect->QRectF api
change is picked up properly by cmake, QWT_VERSION does not work directly in
the code generator.

The colormap store needs to store colors and create new colormaps as the
new api takes ownership of a pointer instead of copying.

git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@37352 ac9c0da5-131f-0410-9009-a9c195f183e6