Added basic target addition functionality to hif_editimlist.
This reuses much of the same code from hif_makeimlist, and
is therefore not yet tuned to the specific requirements
motivation the creation of a new task.  Future commits
are to include the addition of new parameters, possibly
removed heuristics and a possible delete target option.

The reading from a file code is also not yet committed.


git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@39275 ac9c0da5-131f-0410-9009-a9c195f183e6