New Development: Yes             JIRA Issue: cas-6688         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes:              Module(s):            Description:
Added performance additions to the VI/VB2 framework to support more
optimal access to the pointing table. The pointing table has no
enforced structure so it can be easy to access it in a suboptimal
fashion.  Added a caching mechanism that will keep a window of
pointing information around.  Tried simply caching the directions as
measures but that turns out to be horribly expensive computationally,
so it the cache loads the window into memory, keeping the row number
in the cache; it actually loads the direction as a measure when the
pointing is explicitly requested and then keeps that measure cached.

It appears to signficantly improve the performance over the benchmark
using the VI/VB1 framework.  However, it may well have problems with
different pointing tables which will manifest as horrible performance;
since there are no firm requiremnts on how the data in the POINTING
subtable are ordered, it is always possible that some arrangements
will lead to seriously suboptimal performance.  It would probably be a
good idea to have the filler impose some useful order on the subtable
when it loads it; it could also add keywords to the table which would
make it easier to for the accessing logic to determine a good approach
to traversing the table.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@39426 ac9c0da5-131f-0410-9009-a9c195f183e6