Updated hif_editimlist to allow multiple images from field parameter
This is something of a workaround to allow testing of VLASS imaging
by setting a string of field ids, or field names.  hif_editimlist will create
a single target entry for each field in the list.

The field parameter must be a list of either field names as strings
or field ids as integers.  The two types can be combined into a single
list.

For now the imagename parameter is not used.  Instead, the image
name is the same as the fieldname.

git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@39731 ac9c0da5-131f-0410-9009-a9c195f183e6