New Development: Yes             JIRA Issue: cas-6688         Ready for Test: Yes      Interface Changes: What Interface Changed:          Test Programs:   Put in Release Notes:              Module(s):            Description:
Modified the Vi2 framework's frequency selection interface to allow doubly
specifying the desired channels.  The initial selection is made using
channels or an MSSelection.  That selection is then refined by applying
a second frequecy-range selection; the result is that only channels
selected by both methods are ultimately brought out to the VB2.
	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@39881 ac9c0da5-131f-0410-9009-a9c195f183e6