CAS-9631: Exposed threshold_nsigma as editimlist parameter.
This is a multiplier on the robust, full image, rms to determine
the threshold value for each tclean call.  The residual rms, peak
rms, nsigma, peak/rms and threshold are in the log file.  The 
peak/rms is reported in the makeimages web log.

The threhsold_nsigma value default is 4.0.  I also fixed a bug that
was preventing an override of niter.

git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@40437 ac9c0da5-131f-0410-9009-a9c195f183e6