CAS-11038, CAS-9635: hif_editimlist update for VLASS SEIP.
These updates are for the VLASS single epoch imaging pipeline (SEIP).
When the mode is set to VLASS-QL or VLASS-SE-CONT there is a single,
combined image made for the spectral windows and polarizations.
However, when imaging_mode = VLASS-SE-CUBE, the editimlist will create
a target entry for each spw / polarization pair.

git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@41243 ac9c0da5-131f-0410-9009-a9c195f183e6